src/training/federated_train.py
"""Federated training orchestration."""

import flwr as fl
from typing import List, Dict
import yaml


def run_federated_training(
    clients: List,
    config_path: str = "src/config/federated_config.yaml"
):
    """
    Run federated training across all clients.
    
    Args:
        clients: List of federated learning clients
        config_path: Path to federated configuration
    """
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    
    # TODO: Implement federated training loop
    pass


def train_single_round(clients: List, round_num: int) -> Dict:
    """
    Train a single federated round.
    
    Args:
        clients: List of clients
        round_num: Current round number
        
    Returns:
        Round metrics
    """
    # TODO: Implement single round training
    pass
