requirements.txt
# FED-MED Requirements
# Optimized for 32GB VRAM training

# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Data Processing
datasets>=2.16.0
pandas>=2.0.0
numpy>=1.24.0

# Training & Optimization
tensorboard>=2.15.0
wandb>=0.16.0
scipy>=1.11.0

# Federated Learning
# Note: You may need to implement custom FL or use frameworks like:
# flower>=1.6.0  # Flower federated learning framework (optional)

# Configuration
pyyaml>=6.0.1
omegaconf>=2.3.0

# Safety & Validation
regex>=2023.12.0

# Utilities
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Demo (Optional)
streamlit>=1.29.0
gradio>=4.13.0

# Development Tools
ipykernel>=6.27.0
jupyter>=1.0.0
black>=23.12.0
flake8>=7.0.0

# System Monitoring
nvidia-ml-py3>=7.352.0
psutil>=5.9.0
